tasks

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,29 +1,24 @@
 #!/usr/bin/python3
+"""Script deletes all State objects with a name containing the letter a
+Takes three arguments
+    mysql username
+    mysql password
+    database name
+Connects to host localhost and default port (3306)
 """
-This script deletes all State objects
-with a name containing the letter `a`
-from the database `hbtn_0e_6_usa`.
-"""
-
-from sys import argv
-from model_state import Base, State
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-
 if __name__ == "__main__":
-    """
-    Deletes State objects on the database.
-    """
-
-    db_uri = 'mysql+mysqldb://{}:{}@localhost:3306/{}'.format(
-        argv[1], argv[2], argv[3])
-    engine = create_engine(db_uri)
-    Session = sessionmaker(bind=engine)
-
-    session = Session()
-
-    for instance in session.query(State).filter(State.name.contains('a')):
-        session.delete(instance)
-
-    session.commit()
+    from sqlalchemy import (create_engine)
+    from sqlalchemy.orm import sessionmaker
+    from model_state import Base, State
+    from sys import argv
+    Session = sessionmaker()
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.format(
+        argv[1], argv[2], argv[3]), pool_pre_ping=True)
+    session = Session(bind=engine)
+    Base.metadata.create_all(engine)
+    instances = session.query(State).filter(State.name.like('%a%')).all()
+    if instances:
+        for instance in instances:
+            session.delete(instance)
+        session.commit()
     session.close()
